move wallet button hooks from react to base-ui package
