+++ packages/ui/ant-design/src/BaseWalletConnectionButton.tsx
@@ -1,8 +1,7 @@
+import type { WalletName } from '@solana/wallet-adapter-base';
 import type { ButtonProps } from 'antd';
 import { Button } from 'antd';
 import React from 'react';
-
-import type { WalletName } from '@solana/wallet-adapter-base';
 import { WalletIcon } from './WalletIcon.js';
 
 type Props = ButtonProps & {
