+++ packages/ui/base-ui/README.md
@@ -1,0 +1,5 @@
+# `@solana/wallet-adapter-base-ui`
+
+<!-- @TODO -->
+
+Coming soon.
