+++ packages/ui/base-ui/tsconfig.cjs.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../../tsconfig.cjs.json",
+    "include": ["src"],
+    "compilerOptions": {
+        "outDir": "lib/cjs",
+        "jsx": "react"
+    }
+}
