+++ packages/ui/base-ui/tsconfig.esm.json
@@ -1,0 +1,9 @@
+{
+    "extends": "../../../tsconfig.esm.json",
+    "include": ["src"],
+    "compilerOptions": {
+        "outDir": "lib/esm",
+        "declarationDir": "lib/types",
+        "jsx": "react"
+    }
+}
