+++ packages/ui/material-ui/src/BaseWalletConnectButton.tsx
@@ -1,7 +1,7 @@
 import type { ButtonProps } from '@mui/material';
-import React from 'react';
 
-import { useWalletConnectButton } from '@solana/wallet-adapter-react';
+import { useWalletConnectButton } from '@solana/wallet-adapter-base-ui';
+import React from 'react';
 import { BaseWalletConnectionButton } from './BaseWalletConnectionButton.js';
 
 type Props = ButtonProps & {
