+++ packages/ui/material-ui/src/BaseWalletDisconnectButton.tsx
@@ -1,5 +1,5 @@
-import { useWalletDisconnectButton } from '@solana/wallet-adapter-react';
 import type { ButtonProps } from '@mui/material';
+import { useWalletDisconnectButton } from '@solana/wallet-adapter-base-ui';
 import React from 'react';
 import { BaseWalletConnectionButton } from './BaseWalletConnectionButton.js';
