+++ packages/ui/material-ui/src/BaseWalletMultiButton.tsx
@@ -8,9 +8,9 @@ import {
 } from '@mui/icons-material';
 import type { ButtonProps, Theme } from '@mui/material';
 import { Collapse, Fade, ListItemIcon, Menu, MenuItem, styled } from '@mui/material';
-import React, { useMemo, useState } from 'react';
 
-import { useWalletMultiButton } from '@solana/wallet-adapter-react';
+import { useWalletMultiButton } from '@solana/wallet-adapter-base-ui';
+import React, { useMemo, useState } from 'react';
 import { BaseWalletConnectionButton } from './BaseWalletConnectionButton.js';
 import { useWalletDialog } from './useWalletDialog.js';
