+++ packages/ui/react-ui/src/BaseWalletConnectButton.tsx
@@ -1,4 +1,4 @@
-import { useWalletConnectButton } from '@solana/wallet-adapter-react';
+import { useWalletConnectButton } from '@solana/wallet-adapter-base-ui';
 import React from 'react';
 import { BaseWalletConnectionButton } from './BaseWalletConnectionButton.js';
 import type { ButtonProps } from './Button.js';
