+++ packages/ui/react-ui/src/BaseWalletDisconnectButton.tsx
@@ -1,4 +1,4 @@
-import { useWalletDisconnectButton } from '@solana/wallet-adapter-react';
+import { useWalletDisconnectButton } from '@solana/wallet-adapter-base-ui';
 import React from 'react';
 import { BaseWalletConnectionButton } from './BaseWalletConnectionButton.js';
 import type { ButtonProps } from './Button.js';
