+++ packages/ui/react-ui/src/BaseWalletMultiButton.tsx
@@ -1,4 +1,4 @@
-import { useWalletMultiButton } from '@solana/wallet-adapter-react';
+import { useWalletMultiButton } from '@solana/wallet-adapter-base-ui';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { BaseWalletConnectionButton } from './BaseWalletConnectionButton.js';
 import type { ButtonProps } from './Button.js';
