+++ tsconfig.all.json
@@ -26,6 +26,9 @@
             "path": "./packages/ui/ant-design/tsconfig.json"
         },
         {
+            "path": "./packages/ui/base-ui/tsconfig.json"
+        },
+        {
             "path": "./packages/ui/material-ui/tsconfig.json"
         },
         {
